utils: drop None-valued fields from _init_kwargs
The motivation for this change is partly to enable passing
the spider parameters as kwargs to 'action functions' which
can the fail if the parameters don't match the specification
(i.e. the signature) of the action. To enable this method
of verification we need to drop None-valued parameters.

We still leave them in _all_kwargs.

Signed-off-by: Szymon Łopaciuk <szymon.lopaciuk@cern.ch>