use self.logger in all of the spiders (#183)
This is to unify logging mechanisms used in all of the
spiders. Previously some spiders used `self.log`.

In some cases (spiders.common.*) spiders create the
logger independently via `logging.getLogger`. I am
leaving this, as it provides additional debugging
information to know that the message comes from
the superclass.

Signed-off-by: Szymon Łopaciuk <szymon.lopaciuk@cern.ch>