desy: make scraping error JSON serializable
This is a quick fix to avoid raising an exception when spawning the
Celery task that creates the workflow object, as the JSON serializer
used by Celery doesn't know how to serialize exception objects.

A more permanent fix might involve rearchitecting the HEPCrawl items
so that they include the ``errors``, ``traceback`` and perhaps other
keys, in order for inspire-crawler to consume a uniform protocol.

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>